services/web-ui/requirements.txt
flask==3.0.3
gunicorn==22.0.0
requests==2.31.0
kafka-python==2.0.2
prometheus-client==0.20.0
